feat: extract component from FacilityRegistration.vue
